examples/latin_lineation_stub/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "latin-lineation-stub"
version = "0.1.0"
description = "Example mask lineation plugin for transcriber-shell (not a real model)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["numpy>=1.26", "pillow>=10.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/latin_lineation_stub"]
